requirements.txt
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
decorator==5.1.1
et-xmlfile==1.1.0
filelock==3.4.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipython==7.16.3
ipython-genutils==0.2.0
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.17.2
jmespath==0.10.0
lxml==4.8.0
openpyxl==3.0.9
parsel==1.6.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
Protego==0.2.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.11.2
pyOpenSSL==22.0.0
python-decouple==3.6
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
Scrapy==2.6.1
scrapy-proxies==0.4
service-identity==21.1.0
six==1.16.0
tldextract==3.1.2
traitlets==4.3.3
Twisted==22.1.0
typing-extensions==4.1.1
urllib3==1.26.8
w3lib==1.22.0
wcwidth==0.2.5
zope.interface==5.4.0
